Squashed commit of the following:
commit f25399b006e5b15b46e4258546d80e1d006e1bc1
Author: U-bwin7\bernard <bernard@bwin7.(none)>
Date:   Thu May 13 11:41:01 2010 -0400

    VS2010 port

+++ cpp/src/FreezeScript/transformdb.cpp
@@ -686,7 +686,7 @@ run(const Ice::StringSeq& originalArgs, const Ice::CommunicatorPtr& communicator
         while(true)
         {
             in.read(buff, 1024);
-            descriptors.append(buff, in.gcount());
+            descriptors.append(buff, static_cast<size_t>(in.gcount()));
             if(in.gcount() < 1024)
             {
                 break;
